Several changes
1. Return size_t for object id() methods

2. Moved dynamic linking logic from coek/solvers to coek/util

3. Reworked tictoc API.  Standardized on system time, and
   optimized the behavior of tic() and toc() when no string is
   used.

4. Added a SolverResults class, and reworked the solver API
   to return a SolverResults object with suitable information.

+++ lib/coek/coek/api/expression.cpp
@@ -240,7 +240,7 @@ Variable& Variable::name(const std::string& name)
 
 std::string Variable::name() const { return repn->get_name(); }
 
-unsigned int Variable::id() const { return repn->index; }
+size_t Variable::id() const { return repn->index; }
 
 Variable& Variable::within(VariableTypes vtype)
 {
